refactor: migrate to Firebase KTX in the main modules
