+++ app/src/fdroid/java/com/geeksville/mesh/analytics/NopAnalytics.kt
@@ -3,11 +3,19 @@ package com.geeksville.mesh.analytics
 import android.content.Context
 import com.geeksville.mesh.android.Logging
 
+class DataPair(val name: String, valueIn: Any?) {
+    val value = valueIn ?: "null"
+
+    /// An accumulating firebase event - only one allowed per event
+    constructor(d: Double) : this("BOGUS", d)
+    constructor(d: Int) : this("BOGUS", d)
+}
+
 /**
  * Implement our analytics API using Firebase Analytics
  */
 @Suppress("UNUSED_PARAMETER")
-class GoogleAnalytics(context: Context) : AnalyticsProvider, Logging {
+class NopAnalytics(context: Context) : AnalyticsProvider, Logging {
 
     init {
     }
