+++ app/src/fdroid/java/com/geeksville/mesh/android/GeeksvilleApplication.kt
@@ -1,10 +1,8 @@
 package com.geeksville.mesh.android
 
-import android.app.Activity
 import android.app.Application
 import android.content.Context
 import android.content.SharedPreferences
-import android.os.Bundle
 import android.provider.Settings
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.edit
@@ -17,35 +15,6 @@ open class GeeksvilleApplication : Application(), Logging {
 
     companion object {
         lateinit var analytics: AnalyticsProvider
-        var currentActivity: Activity? = null
-        private val backstack = mutableListOf<Activity>()
-    }
-
-    private val lifecycleCallbacks = object : ActivityLifecycleCallbacks {
-        override fun onActivityPaused(activity: Activity) {
-        }
-
-        override fun onActivityStarted(activity: Activity) {
-        }
-
-        override fun onActivityDestroyed(activity: Activity) {
-            if (backstack.contains(activity)) backstack.remove(activity)
-            currentActivity = backstack.lastOrNull()
-        }
-
-        override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {
-        }
-
-        override fun onActivityStopped(activity: Activity) {
-        }
-
-        override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
-            backstack.add(activity)
-            currentActivity = backstack.lastOrNull()
-        }
-
-        override fun onActivityResumed(activity: Activity) {
-        }
     }
 
     /// Are we running inside the testlab?
@@ -81,12 +50,8 @@ open class GeeksvilleApplication : Application(), Logging {
     override fun onCreate() {
         super.onCreate()
 
-        val googleAnalytics = com.geeksville.mesh.analytics.GoogleAnalytics(this)
-        analytics = googleAnalytics
-
-        // Set analytics per prefs
-        isAnalyticsAllowed = isAnalyticsAllowed
-
-        registerActivityLifecycleCallbacks(lifecycleCallbacks)
+        val nopAnalytics = com.geeksville.mesh.analytics.NopAnalytics(this)
+        analytics = nopAnalytics
+        isAnalyticsAllowed = false
     }
 }
